sim(64-bit): make sample-path pointer math 64-bit-clean
The sample reader/recorder computed byte offsets by casting char* to uint32_t and
subtracting, which truncates on a 64-bit host. Express the intent directly: native
pointer subtraction (ptrdiff_t) for cluster byte offsets, char* arithmetic for
byte-offset-into-buffer, and uintptr_t only where a char* deliberately holds a small
integer offset (the null-cluster silence path). Values are unchanged; the code is
clearer than the integer-cast form it replaces.

+++ src/deluge/model/sample/sample_recorder.cpp
@@ -606,7 +606,7 @@ Error SampleRecorder::finalizeRecordedFile() {
 
 	// In the very rare case where we've already got between 1 and 5 bytes overhanging the end of our current cluster,
 	// we need to allocate a new one right now
-	int32_t bytesTilClusterEnd = (uint32_t)clusterEndPos - (uint32_t)writePos;
+	int32_t bytesTilClusterEnd = clusterEndPos - writePos;
 	if (bytesTilClusterEnd < 0) {
 		Error error = createNextCluster();
 
@@ -629,7 +629,7 @@ Error SampleRecorder::finalizeRecordedFile() {
 	// created, cos or RAM or file size limit.)
 	if (currentRecordCluster) {
 
-		int32_t bytesToWrite = (uint32_t)writePos - (uint32_t)currentRecordCluster->data;
+		int32_t bytesToWrite = writePos - currentRecordCluster->data;
 		if (bytesToWrite > 0) { // Will always be true
 			Error error = writeCluster(currentRecordClusterIndex, bytesToWrite);
 			if (error != Error::NONE) {
@@ -856,7 +856,7 @@ Error SampleRecorder::createNextCluster() {
 	if (currentRecordClusterIndex >= (1 << (MAX_FILE_SIZE_MAGNITUDE - Cluster::size_magnitude))) {
 
 		// See if we actually already had any bytes to write into that new cluster we can't have...
-		int32_t bytesTilClusterEnd = (uint32_t)clusterEndPos - (uint32_t)writePos;
+		int32_t bytesTilClusterEnd = clusterEndPos - writePos;
 		if (bytesTilClusterEnd < 0) {
 			numSamplesCaptured--;
 			writePos -= (int32_t)recordingNumChannels * 3;
@@ -894,7 +894,7 @@ Error SampleRecorder::createNextCluster() {
 	memcpy(currentRecordCluster->data, &oldRecordCluster->data[Cluster::size],
 	       5); // 5 is the max number of bytes we could have overshot
 
-	int32_t bytesOvershot = (uint32_t)writePos - (uint32_t)clusterEndPos;
+	int32_t bytesOvershot = writePos - clusterEndPos;
 
 	currentRecordCluster->loaded =
 	    true; // I think this is ok - mark it as loaded even though we're yet to record into it
@@ -952,7 +952,7 @@ doFinishCapturing:
 			int32_t bytesPerSample = recordingNumChannels * 3;
 			int32_t bytesWeWantToWrite = numSamplesThisCycle * bytesPerSample;
 
-			int32_t bytesTilClusterEnd = (uint32_t)clusterEndPos - (uint32_t)writePos;
+			int32_t bytesTilClusterEnd = clusterEndPos - writePos;
 
 			// If we need a new cluster right now...
 			if (bytesTilClusterEnd <= 0) {
@@ -966,7 +966,7 @@ doFinishCapturing:
 					return;
 				}
 
-				bytesTilClusterEnd = (uint32_t)clusterEndPos - (uint32_t)writePos; // Recalculate it
+				bytesTilClusterEnd = clusterEndPos - writePos; // Recalculate it
 			}
 
 			if (bytesTilClusterEnd <= bytesWeWantToWrite - bytesPerSample) {
@@ -1304,7 +1304,7 @@ Error SampleRecorder::alterFile(MonitoringAction action, int32_t lshiftAmount, u
 
 		// If need to advance write-head past the end of a cluster, then we'll write that current cluster to disk
 		// and carry on
-		int32_t writeOvershot = (uint32_t)writePos - (uint32_t)&currentWriteCluster->data[Cluster::size];
+		int32_t writeOvershot = writePos - &currentWriteCluster->data[Cluster::size];
 		if (writeOvershot >= 0) {
 
 			// If reached very end of file, break
@@ -1413,7 +1413,7 @@ writeFailed:
 
 			D_PRINTLN("read advance");
 
-			int32_t overshot = (uint32_t)readPos - (uint32_t)&currentReadCluster->data[Cluster::size];
+			int32_t overshot = readPos - &currentReadCluster->data[Cluster::size];
 
 			// Some bug-hunting
 			if (!currentReadCluster->numReasonsHeldBySampleRecorder) {
@@ -1476,7 +1476,7 @@ writeFailed:
 
 	currentWriteCluster->loaded = true;
 
-	uint32_t bytesToWriteFinalCluster = (uint32_t)writePos - (uint32_t)currentWriteCluster->data;
+	uint32_t bytesToWriteFinalCluster = writePos - currentWriteCluster->data;
 
 	if (bytesToWriteFinalCluster) { // If there is in fact anything to flush out to the file / card...
